requirements.txt
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flask-cors==4.0.1
Flask-PyMongo==3.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
PyJWT==2.10.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soupsieve==2.6
threadpoolctl==3.6.0
tldextract==5.1.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
whois==1.20240129.2
xgboost==3.0.0
